+++ demo-html/main.tsx
@@ -1,0 +1,4 @@
+
+/*tslint:disable:no-import-side-effect*/
+import './components/counter';
+import './components/xkcd';