+++ demo/components/store.ts
@@ -1,72 +1,68 @@
 function _random(max) {
-    return Math.round(Math.random()*1000)%max;
+  return Math.round(Math.random()*1000)%max;
 }
 
-export default class Store {
-    data = [];
-    backup = null;
-    selected = null;
-    id = 1;
-    constructor() {
-    }
-    buildData(count = 1000) {
-        var adjectives = ["pretty", "large", "big", "small", "tall", "short", "long", "handsome", "plain", "quaint", "clean", "elegant", "easy", "angry", "crazy", "helpful", "mushy", "odd", "unsightly", "adorable", "important", "inexpensive", "cheap", "expensive", "fancy"];
-        var colours = ["red", "yellow", "blue", "green", "pink", "brown", "purple", "brown", "white", "black", "orange"];
-        var nouns = ["table", "chair", "house", "bbq", "desk", "car", "pony", "cookie", "sandwich", "burger", "pizza", "mouse", "keyboard"];
-        var data = [];
-        for (var i = 0; i < count; i++)
-            data.push({id: this.id++, label: adjectives[_random(adjectives.length)] + " " + colours[_random(colours.length)] + " " + nouns[_random(nouns.length)] });
-        return data;
-    }
-    updateData(mod = 10) {
-        for (let i=0;i<this.data.length;i+=10) {
-            this.data[i].label += ' !!!';
-            // this.data[i] = Object.assign({}, this.data[i], {label: this.data[i].label +' !!!'});
-        }
-    }
-    delete(id) {
-        // const idx = this.data.findIndex(d => d.id==id);
-        this.data = this.data.filter((e, i) => e.id != id);
-        if (id === this.selected) this.selected = null;
-    }
-    run() {
-        this.data = this.buildData();
-        this.selected = null;
-    }
-    add() {
-        this.data = this.data.concat(this.buildData(1000));
-        this.selected = null;
-    }
-    update() {
-        this.updateData();
-    }
-    select(id) {
-        this.selected = id;
-    }
-    hideAll() {
-        this.backup = this.data;
-        this.data = [];
-        this.selected = null;
-    }
-    showAll() {
-        this.data = this.backup;
-        this.backup = null;
-        this.selected = null;
-    }
-    runLots() {
-        this.data = this.buildData(10000);
-        this.selected = null;
-    }
-    clear() {
-        this.data = [];
-        this.selected = null;
-    }
-    swapRows() {
-        if (this.data.length > 4) {
-            var idx = this.data.length - 2;
-            var a = this.data[1];
-            this.data[1] = this.data[idx];
-            this.data[idx] = a;
-        }
-    }
+const adjectives = ["pretty", "large", "big", "small", "tall", "short", "long", "handsome", "plain", "quaint", "clean", "elegant", "easy", "angry", "crazy", "helpful", "mushy", "odd", "unsightly", "adorable", "important", "inexpensive", "cheap", "expensive", "fancy"];
+const colours = ["red", "yellow", "blue", "green", "pink", "brown", "purple", "brown", "white", "black", "orange"];
+const nouns = ["table", "chair", "house", "bbq", "desk", "car", "pony", "cookie", "sandwich", "burger", "pizza", "mouse", "keyboard"];
+
+let id = 1
+function buildData(count) {
+  return new Array(count).fill(0).map(_ => ({
+    id: id++,
+    label: `${adjectives[_random(adjectives.length)]} ${colours[_random(colours.length)]} ${nouns[_random(nouns.length)]}`
+  }))
 }
+
+export default {
+  run: () => ({
+    data: buildData(1000),
+    selected: null
+  }),
+
+  add: state => ({
+    data: state.data.concat(buildData(1000)),
+    selected: state.selected,
+  }),
+
+  runLots: () => ({
+    data: buildData(10000),
+    selected: null
+  }),
+
+  clear: () => ({
+    data: [],
+    selected: null
+  }),
+
+  update: state => ({
+    data: state.data.map((d, i) => {
+      if (i % 10 === 0) {
+        d.label = `${d.label} !!!`
+      }
+      return d
+    }),
+    selected: state.selected
+  }),
+
+  swapRows: state => {
+    if (state.data.length > 4) {
+      const idx = state.data.length - 2;
+      const a = state.data[1];
+      state.data[1] = state.data[idx];
+      state.data[idx] = a;
+    }
+    return state;
+  },
+
+  select: (state, id) => ({
+    data: state.data,
+    selected: id
+  }),
+
+  delete: (state, id) => ({
+    selected: id === state.selected ? null : state.selected,
+    data: state.data.filter(d => d.id != id)
+  })
+}
+