+++ src/vdom-patch.ts
@@ -4,13 +4,14 @@ enum CHANGES {
 }
 
 function patch(vdom1, vdom2) {
+  if (!vdom1?.length || !vdom2?.length) return;
   const old_len = vdom1.length;
   const new_len = vdom2.length;
   const len = Math.min(old_len, new_len);
   for (let i = 0; i < len; i++) {
     const node1 = vdom1[i];
     const node2 = vdom2[i];
-    if (typeof node2 !== 'object') continue;
+    if (!node2 || typeof node2 !== 'object') continue;
     if (_areEquals(node1, node2)) {
       node2._op = CHANGES.NO_ATTR_CHANGE | CHANGES.NO_TREE_CHANGE;
     } else {